+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Modal autofocus was not working due to new dialog
 - Center modal title on mobile
+- Modal may show second scrollbar in some cases of y-overflow
 
 ### Changed
 - Initially hide main container scrollbar on small devices